+++ sites/kit.svelte.dev/src/lib/docs/TSToggle.svelte
@@ -1,5 +1,5 @@
 <script>
-	import Checkbox from './Checkbox.svelte';
+	import ToggleButton from './ToggleButton.svelte';
 
 	let checked = prefers_ts();
 	$: toggle(checked);
@@ -22,10 +22,9 @@
 	}
 </script>
 
-<!-- svelte-ignore a11y-label-has-associated-control -->
-<label class="input-output-toggle">
-	JavaScript <Checkbox bind:checked /> TypeScript
-</label>
+<div class="input-output-toggle">
+	<span aria-hidden="true">JavaScript</span> <ToggleButton bind:pressed={checked} label="TypeScript code examples" /> <span aria-hidden="true">TypeScript</span>
+</div>
 
 <style>
 	.input-output-toggle {
@@ -35,8 +34,17 @@
 		user-select: none;
 		align-items: center;
 		width: 100%;
-		height: 42px;
+		height: var(--ts-toggle-height);
 		z-index: 2;
-		padding: 0 3.2rem;
+		padding: 0 var(--sk-page-padding-side);
+		margin: 0 auto;
+	}
+
+	@media (min-width: 832px) {
+		.input-output-toggle {
+			padding-left: 3.2rem;
+			width: var(--sidebar-menu-width);
+			margin: 0 0 0 auto;
+		}
 	}
 </style>