+++ sites/kit.svelte.dev/src/lib/docs/server/index.js
@@ -72,6 +72,7 @@ export async function read_file(file) {
 	const { metadata, body } = extract_frontmatter(markdown);
 
 	const { content, sections } = parse({
+		file,
 		body: generate_ts_from_js(body),
 		code: (source, language, current) => {
 			const hash = createHash('sha256');
@@ -107,10 +108,10 @@ export async function read_file(file) {
 			let version_class = '';
 			if (language === 'generated-ts' || language === 'generated-svelte') {
 				language = language.replace('generated-', '');
-				version_class = ' ts-version';
+				version_class = 'ts-version';
 			} else if (language === 'original-js' || language === 'original-svelte') {
 				language = language.replace('original-', '');
-				version_class = ' js-version';
+				version_class = 'js-version';
 			}
 
 			if (language === 'dts') {
@@ -221,9 +222,13 @@ export async function read_file(file) {
 				html = `<pre class='language-${plang}'><code>${highlighted}</code></pre>`;
 			}
 
-			html = `<div class="code-block${version_class} ${options.style ?? ''}">${
-				options.file ? `<h5>${options.file}</h5>` : ''
-			}${html}</div>`;
+			if (options.file) {
+				html = html.replace('<pre', `<pre data-file="${options.file}"`);
+			}
+
+			if (version_class) {
+				html = html.replace(/class=('|")/, `class=$1${version_class} `);
+			}
 
 			type_regex.lastIndex = 0;
 
@@ -282,12 +287,13 @@ export async function read_file(file) {
 
 /**
  * @param {{
+ *   file: string;
  *   body: string;
  *   code: (source: string, language: string, current: string) => string;
  *   codespan: (source: string) => string;
  * }} opts
  */
-function parse({ body, code, codespan }) {
+function parse({ file, body, code, codespan }) {
 	const headings = [];
 
 	/** @type {import('./types').Section[]} */
@@ -322,7 +328,7 @@ function parse({ body, code, codespan }) {
 
 			const slug = headings.filter(Boolean).join('-');
 
-			if (level === 3) {
+			if (level === 2) {
 				section = {
 					title,
 					slug,
@@ -330,16 +336,16 @@ function parse({ body, code, codespan }) {
 				};
 
 				sections.push(section);
-			} else if (level === 4) {
-				section.sections.push({
+			} else if (level === 3) {
+				(section?.sections ?? sections).push({
 					title,
 					slug
 				});
 			} else {
-				// throw new Error(`Unexpected <h${level}> in ${file}`);
+				throw new Error(`Unexpected <h${level}> in ${file}`);
 			}
 
-			return `<h${level} id="${slug}">${html}<a href="#${slug}" class="anchor"><span class="visually-hidden">permalink</span></a></h${level}>`;
+			return `<h${level} id="${slug}">${html}<a href="#${slug}" class="permalink"><span class="visually-hidden">permalink</span></a></h${level}>`;
 		},
 		code: (source, language) => code(source, language, current),
 		codespan